requirements.txt
asgiref==3.3.1
Django==3.1.7
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
Pillow==8.2.0
PyJWT==1.7.1
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
